Separate tsconfig.json files, remove `dist` from the repo
