+++ src/typescript/sdk/tests/tsconfig.json
@@ -1,7 +1,9 @@
 {
   "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
   "include": [
-    "**/*.ts",
-    "../src/utils/aptos-client.ts"
+    "./**/*"
   ]
 }
