+++ src/typescript/sdk/tsconfig.json
@@ -13,10 +13,9 @@
     "module": "esnext",
     "moduleDetection": "force",
     "moduleResolution": "bundler",
-    "noEmit": false,
+    "noEmit": true,
     "noImplicitAny": true,
     "noImplicitThis": true,
-    "outDir": "./dist",
     "preserveWatchOutput": true,
     "resolveJsonModule": true,
     "skipLibCheck": true,
@@ -25,6 +24,6 @@
     "target": "esnext",
     "types": ["node", "jest"]
   },
-  "exclude": ["node_modules", "**/*.test.ts", "**/*.spec.ts", "tests"],
+  "exclude": ["node_modules"],
   "include": ["src"]
 }
