+++ src/typescript/sdk/tsconfig.prod.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": false,
+    "outDir": "./dist",
+  },
+  "exclude": ["node_modules", "**/*.test.ts", "**/*.spec.ts", "tests/e2e/**/*", "tests/unit/**/*"],
+  "include": ["src", "dist"]
+}
